+++ api-server/src/events/events.controller.ts
@@ -160,29 +160,6 @@ export class EventsController {
       });
   }
 
-  // TODO - Move to events.authenticated.controller
-  @Get('non-verified')
-  @UseGuards(AuthGuard)
-  @ApiOperation({
-    summary: 'Get events that have not yet been verified (admin only)',
-  })
-  @ApiResponse({ status: 403, description: 'Forbidden' })
-  @ApiBearerAuth()
-  getAllNonVerified(
-    @Query('embed') embed: string[] | string = [],
-    @Query('tags') tags: string[] | string = [],
-  ): Promise<EventsResponse> {
-    const findOptions = {
-      ...getOptionsForEventsServiceFromEmbedsQueryParam(embed),
-      where: { verified: false },
-    };
-
-    return this.eventsService
-      .findAll(findOptions)
-      .then((events) => events.map(eventModelToEventDTO))
-      .then((events) => new EventsResponse({ events }));
-  }
-
   @Get('/:id')
   @ApiOperation({ summary: 'Get single event by id' })
   @ApiResponse({
@@ -206,27 +183,6 @@ export class EventsController {
       .then(eventModelToEventDTO)
       .then((event) => removeSensitiveDataForNonAdmins(request, event))
       .then((event: EventDTO) => ({ event, status: 'success' }));
-  }
-
-  // TODO - Move to events.authenticated.controller
-  @Get()
-  @UseGuards(AuthGuard)
-  @ApiOperation({ summary: 'Get events, both verified and non (admin only)' })
-  @ApiResponse({ status: 403, description: 'Forbidden' })
-  @ApiBearerAuth()
-  getAll(
-    @Query('embed') embed: string[] | string = [],
-    @Query('tags') tags: string[] | string = [],
-  ): Promise<EventsResponse> {
-    const findOptions = {
-      ...getOptionsForEventsServiceFromEmbedsQueryParam(embed),
-      where: getCommonQueryTermsForEvents(null, tags),
-    };
-
-    return this.eventsService
-      .findAll(findOptions)
-      .then((events) => events.map(eventModelToEventDTO))
-      .then((events) => new EventsResponse({ events }));
   }
 
   @Post()
